+++ sdk/test/metrics/sync_metric_storage_up_down_counter_test.cc
@@ -99,6 +99,10 @@ TEST_P(WritableMetricStorageTestFixture, LongUpDownCounterSumAggregation)
   count_attributes = 0;
   storage.Collect(collector.get(), collectors, sdk_start_ts, collection_ts,
                   [&](const MetricData data) {
+                    if (temporality == AggregationTemporality::kCumulative)
+                    {
+                      EXPECT_EQ(data.start_ts, sdk_start_ts);
+                    }
                     for (auto data_attr : data.point_data_attr_)
                     {
                       auto sum_data = opentelemetry::nostd::get<SumPointData>(data_attr.point_data);
@@ -246,6 +250,10 @@ TEST_P(WritableMetricStorageTestFixture, DoubleUpDownCounterSumAggregation)
   count_attributes = 0;
   storage.Collect(collector.get(), collectors, sdk_start_ts, collection_ts,
                   [&](const MetricData data) {
+                    if (temporality == AggregationTemporality::kCumulative)
+                    {
+                      EXPECT_EQ(data.start_ts, sdk_start_ts);
+                    }
                     for (auto data_attr : data.point_data_attr_)
                     {
                       auto sum_data = opentelemetry::nostd::get<SumPointData>(data_attr.point_data);